w01/requirements.txt
Send2Trash==1.8.0
requests==2.28.0
beautifulsoup4==4.11.1
selenium==4.4.3
openpyxl==3.0.10
PyPDF2==2.10.3
python-docx==0.8.11
IMAPClient==2.3.1
pyzmail36==1.0.5
twilio
ezgmail
ezsheets
Pillow==9.2.0
PyAutoGUI
pyperclip
PyInputPlus
python3-xlib==0.15
